Revert "Dashboard page"
