--- fwa_news/lib/Helpers/FlashHelper.dart
@@ -1,117 +1,0 @@
-import 'dart:async';
-
-import 'package:flash/flash.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-
-class FlashHelper {
-  static Completer<BuildContext> _buildCompleter = Completer<BuildContext>();
-
-  static void init(BuildContext context) {
-    if (_buildCompleter?.isCompleted == false) {
-      _buildCompleter.complete(context);
-    }
-  }
-
-  static void dispose() {
-    if (_buildCompleter?.isCompleted == false) {
-      _buildCompleter.completeError(FlutterError('disposed'));
-    }
-    _buildCompleter = Completer<BuildContext>();
-  }
-
-  static TextStyle _titleStyle(BuildContext context, [Color color]) {
-    var theme = Theme.of(context);
-    return (theme.dialogTheme?.titleTextStyle ?? theme.textTheme.title)
-        .copyWith(color: color);
-  }
-
-  static TextStyle _contentStyle(BuildContext context, [Color color]) {
-    var theme = Theme.of(context);
-    return (theme.dialogTheme?.contentTextStyle ?? theme.textTheme.body1)
-        .copyWith(color: color);
-  }
-
-  static Future<T> successBar<T>(
-    BuildContext context, {
-    String title,
-    @required String message,
-    Duration duration = const Duration(seconds: 3),
-  }) {
-    return showFlash<T>(
-      context: context,
-      duration: duration,
-      builder: (_, controller) {
-        return Flash(
-          controller: controller,
-          horizontalDismissDirection: HorizontalDismissDirection.horizontal,
-          backgroundColor: Colors.black87,
-          child: FlashBar(
-            title: title == null
-                ? null
-                : Text(title, style: _titleStyle(context, Colors.white)),
-            message: Text(message, style: _contentStyle(context, Colors.white)),
-            icon: Icon(Icons.check_circle, color: Colors.green[300]),
-            leftBarIndicatorColor: Colors.green[300],
-          ),
-        );
-      },
-    );
-  }
-
-  static Future<T> informationBar<T>(
-    BuildContext context, {
-    String title,
-    @required String message,
-    Duration duration = const Duration(seconds: 3),
-  }) {
-    return showFlash<T>(
-      context: context,
-      duration: duration,
-      builder: (_, controller) {
-        return Flash(
-          controller: controller,
-          horizontalDismissDirection: HorizontalDismissDirection.horizontal,
-          backgroundColor: Colors.black87,
-          child: FlashBar(
-            title: title == null
-                ? null
-                : Text(title, style: _titleStyle(context, Colors.white)),
-            message: Text(message, style: _contentStyle(context, Colors.white)),
-            icon: Icon(Icons.info_outline, color: Colors.blue[300]),
-            leftBarIndicatorColor: Colors.blue[300],
-          ),
-        );
-      },
-    );
-  }
-
-  static Future<T> errorBar<T>(
-    BuildContext context, {
-    String title,
-    @required String message,
-    Duration duration = const Duration(seconds: 3),
-  }) {
-    return showFlash<T>(
-      context: context,
-      duration: duration,
-      builder: (_, controller) {
-        return Flash(
-          controller: controller,
-          horizontalDismissDirection: HorizontalDismissDirection.horizontal,
-          backgroundColor: Colors.black87,
-          child: FlashBar(
-            title: title == null
-                ? null
-                : Text(title, style: _titleStyle(context, Colors.white)),
-            message: Text(message, style: _contentStyle(context, Colors.white)),
-            icon: Icon(Icons.warning, color: Colors.red[300]),
-            leftBarIndicatorColor: Colors.red[300],
-          ),
-        );
-      },
-    );
-  }
-}
-
-typedef ActionCallback = void Function(FlashController controller);
