--- fwa_news/lib/Routes/Routes.dart
@@ -1,6 +1,0 @@
-class Routes
-{
-  static const LOGIN = "/login";
-  static const DASHBOARD = "/dashboard";
-  static const FRAME = "/frame";
-}
