+++ fwa_news/pubspec.yaml
@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -23,13 +23,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  flutter_spinkit: "^2.1.0"
-  shared_preferences: ^0.4.2
-  flash: "^1.1.0"
-  http: ^0.12.0+4
-  transparent_image: ^1.0.0
-  curved_navigation_bar: ^0.3.2
-  sprintf: "^4.0.0"
 
 dev_dependencies:
   flutter_test:
@@ -48,9 +41,9 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-    - assets/images/login_background.png
-    - assets/images/image_placeholder.png
+  # assets:
+  #  - images/a_dot_burr.jpeg
+  #  - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
